+++ tests/test_components/test_types.py
@@ -1,7 +1,7 @@
 """Tests type definitions."""
 import pytest
 import tidy3d as td
-from tidy3d.components.types import ArrayLike, Complex, constrained_array
+from tidy3d.components.types import ArrayLike, Complex, constrained_array, Tuple
 from tidy3d.components.base import Tidy3dBaseModel
 from tidy3d.exceptions import ValidationError
 import numpy as np
@@ -14,6 +14,15 @@ def _test_validate_array_like():
     s = S(f=np.array([[1.0, 2.0, 3.0], [4.0, 5.0, 6.0]]))
     with pytest.raises(pydantic.ValidationError):
         s = S(f=np.array([1.0, 2.0, 3.0]))
+
+    class MyClass(Tidy3dBaseModel):
+        f: constrained_array(ndim=3, shape=(1, 2, 3))
+
+    with pytest.raises(pydantic.ValidationError):
+        _ = MyClass(f=np.ones((2, 2, 3)))
+
+    with pytest.raises(pydantic.ValidationError):
+        _ = MyClass(f=np.ones((1, 2, 3, 4)))
 
 
 def test_schemas():
@@ -35,6 +44,7 @@ def test_array_like():
         c: constrained_array(dtype=float) = None  # must be float-like
         d: constrained_array(ndim=1, dtype=complex) = None  # 1D complex
         e: ArrayLike
+        f: constrained_array(ndim=3, shape=(1, 2, 3)) = None  # must have certain shape
 
     my_obj = MyClass(
         a=1.0 + 2j,
@@ -42,6 +52,7 @@ def test_array_like():
         c=[1, 3.0],
         d=[1.0],
         e=[[[[1.0]]]],
+        f=np.ones((1, 2, 3)),
     )
 
     assert np.all(my_obj.a == [1.0 + 2j])  # scalars converted to list of len 1
@@ -57,6 +68,7 @@ def test_hash():
 
         a: ArrayLike
         b: constrained_array(ndim=1)
+        c: Tuple[ArrayLike, ...]
 
-    c = MyClass(a=[1.0], b=[2.0, 1.0])
+    c = MyClass(a=[1.0], b=[2.0, 1.0], c=([2.0, 1.0]))
     hash(c.json())